+++ laravel/database/migrations/2024_11_02_113824_create_game_invitations_table.php
@@ -11,15 +11,17 @@ return new class extends Migration
      */
     public function up(): void
     {
-        Schema::create('game_invitations', function (Blueprint $table) {
-            $table->id();
-            $table->foreignId('game_id')->unique()->constrained('games')->onDelete('cascade');
-            $table->foreignId('invited_by')->constrained('users');
-            $table->uuid('token')->unique();
-            $table->enum('status', ['active', 'expired'])->default('active');
-            $table->timestamp('expires_at');
-            $table->timestamps();
-        });
+        if (!Schema::hasTable('game_invitations')) {
+            Schema::create('game_invitations', function (Blueprint $table) {
+                $table->id();
+                $table->foreignId('game_id')->unique()->constrained('games')->onDelete('cascade');
+                $table->foreignId('invited_by')->constrained('users');
+                $table->uuid('token')->unique();
+                $table->enum('status', ['active', 'expired'])->default('active');
+                $table->timestamp('expires_at');
+                $table->timestamps();
+            });
+        }
     }
 
     /**
